+++ distribution/ChangeLog.md
@@ -1,4 +1,5 @@
 # 0.2.0 -- yyyy-mm-dd
 
-- Hidden internal modules
-
+- ClickHaskell now have it's own domain name: https://clickhaskell.dev/
+- Improved multithreading connection usage (+ add test)
+- Migration to single module distribution